cancer_classification/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
dicomsdl==0.109.3
executing==2.1.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gradio_client==1.4.3
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.29.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
kagglehub==0.3.12
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
pydicom==3.0.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
python-slugify==8.0.4
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
ruff==0.8.0
safehttpx==0.1.1
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.45.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.3.30
toml==0.10.2
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.3
wrapt==1.16.0
